fix: Allow deploying old commits on default branch
By default, if you do not turn off auto merging, the deployment API will
attempt to merge the default branch into the requested ref. If you are
attempting to deploy an older commit, this will fail.

Since the goal of this tool is to trigger a deployment, not also manage
branches or merges, I am disabling auto merge functionality.