feat: Use current branch if ref not specified
If you have explicitly specified a repository, the assumption is that
you are not in that checkout. This, therefore, requires that you specify
a reference to deploy as this application has no way to determine your
intent.

However, if you are in a repo and do not specify the repository, the
application will assume you mean the current checkout. In this
situation, you can opt not to specify a reference, and the application
will assume you want the current branch.